+++ database/migrations/2021_04_25_175813_create_group_module.php
@@ -54,13 +54,13 @@ class CreateGroupModule extends Migration
                                         [
                                             'name' => 'name',
                                             'type' => 'string',
-                                            'display' => true,
+                                            'visible' => true,
                                             'required' => true,
                                         ],
                                         [
                                             'name' => 'description',
                                             'type' => 'string',
-                                            'display' => true,
+                                            'visible' => true,
                                         ],
                                     ]
                                 ],
@@ -70,9 +70,20 @@ class CreateGroupModule extends Migration
                                     'close' => true,
                                     'fields' => [
                                         [
+                                            'name' => 'workspace',
+                                            'type' => 'entity',
+                                            'visible' => [
+                                                'detail' => true,
+                                                'list' => true
+                                            ],
+                                            'config' => [
+                                                'module' => 'workspace'
+                                            ]
+                                        ],
+                                        [
                                             'name' => 'created_at',
                                             'type' => 'datetime',
-                                            'display' => [
+                                            'visible' => [
                                                 'detail' => true,
                                                 'list' => true
                                             ],
@@ -80,7 +91,7 @@ class CreateGroupModule extends Migration
                                         [
                                             'name' => 'updated_at',
                                             'type' => 'datetime',
-                                            'display' => [
+                                            'visible' => [
                                                 'detail' => true,
                                                 'list' => true
                                             ],